requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
torch
diffusers
huggingface_hub